Fixed REALLY hard to find bug in monitor/garbage collection code.
For some reason the object variable was copied into a local obj variable, which was then added to the safe pointer list. However, the rest of this function doesn't use the new obj variable, but the original object parameter. This means that if the garbage collector is triggered while in this function, the pointer in the new obj variable gets updated to the object's new location, but the original object parameter won't.
The result is a monitor is created for the wrong address. Then the next MONITOREXIT instruction will create yet another monitor and reduce it's count, which will loop around to -1/255. So we end up with two monitors that will never be destroyed again. For the second one (with count 255) this only wastes memory, but for the first we now have an invalid pointer. This will be ok for a while, but after each following run of the garbage collector, the address will be updated by some value determined by whatever happens to be there (which probably won't be a darjeeling heap chunk). This continues until the pointer runs out of the process' address space and we get a segfault.