Added empty hart library.
Not sure yet if this should be a separate lib or part of wkcomm, but it's easier to combine things later than to separate them. (another option would be to use/refactor the code in avr.c)
The idea is that if the posix implementation for wave can use the usb as a uart, all the Zwave protocol code could be moved to common since only the uart communication would be platform specific.