Call the infuser task on the base library before building the vm.
The vm and base libraries depend on each other. Obviously base needs a lot of headers from vm, but vm also needs a number of constants define in base, for example to create Exception objects.
This circular dependency is ugly of course. Maybe we should merge the two later, but for now I'm just calling the "infuser" task in base before building vm. This lets the infuser create the necessary header files without trying to build the C code which would fail. This is all vm needs to build.